fix(web): move workbook Fork button into the metadata header
The Fork action was a standalone button floating above the workbook page,
separated from the title/metadata block the layout renders. Move it into the
Created/Updated/Created by/Version row (right-aligned) in WorkbookLayoutContent,
which already has the workbook and session, and drop it from the page. No
behavior change; the button just reads as part of the header now.
